requirements.txt
Flask==2.3.6
requests==2.31.0
google-genai==0.4.1
